refactor: rate limiter as middleware
